改了nemu异常处理mepc加4的时机。在npc目录下可通过make perf测试性能

+++ npc/csrc/monitor/sdb.c
@@ -7,7 +7,7 @@
 #include <memory.h>
 
 void reg_display();
-void cpu_exec(unsigned n);
+void cpu_exec(unsigned long n);
 
 int batch_mode = 0;
